Modularize javascript logic
