+++ profile.html
@@ -19,7 +19,7 @@
         <li><a href="entries.html">My entries</a></li>
         <li><a href="create.html">Write</a></li>
         <li><a href="profile.html">Profile</a></li>
-        <li class="right"><a href onclick="logout(event)" id="logout">Logout</a></li>
+        <li class="right"><a href id="logout">Logout</a></li>
         <li class="right"><a href="signin.html">Sign in</a></li>
         <li class="right"><a href="signup.html">Sign up</a></li>
     </ul>
