+++ scripts/api.js
@@ -1,0 +1,3 @@
+const API_URL = 'https://mdiary-v1.herokuapp.com/api/v1/';
+
+export const apiUrl = (path) => API_URL + path;
