+++ scripts/app.js
@@ -1,376 +1,1 @@
-const API_URL = 'https://mdiary-v1.herokuapp.com/api/v1/';
-
-const apiUrl = (path) => API_URL + path;
-
-const getToken = () => localStorage.getItem('auth-token');
-
-const setToken = (value) => localStorage.setItem('auth-token', value);
-
-const clearToken = () => localStorage.removeItem('auth-token');
-
-const checkAuth = (uri) => {
-    // the sign up and sign in pages are not protected pages
-    if (uri === 'signup.html' || uri === 'signin.html') {
-        return
-    }
-    if (!getToken()) {
-        // always require login if token is missing
-        window.location.href = 'signin.html';
-    }
-};
-
-const getElement = (id) => document.getElementById(id);
-
-const getValue = (id) => {
-    let input = getElement(id);
-    if (input.value === '') {
-        return null;
-    }
-    return input.value;
-};
-
-const logout = (event) => {
-    event.preventDefault();
-    if (getToken()) {
-        clearToken();
-    }
-    window.location = 'index.html';
-};
-
-const appErrors = $('#app_errors');
-
-const appSuccess = $('#app_success');
-
-const displaySuccess = (message) => {
-    appSuccess.empty();
-    appSuccess.append($('<ul>')
-        .append($('<li>')
-            .html(message)
-        )
-    );
-    appSuccess.show();
-};
-
-const displayError = (message) => {
-    appErrors.empty();
-    appErrors.append($('<ul>')
-        .append($('<li>')
-            .html(message)
-        )
-    );
-    appErrors.show();
-};
-
-const handleErrors = (code, payload) => {
-    appErrors.empty();
-    switch (code) {
-        case 404:
-            return displayError('The entry was not found.');
-        case 422:
-            let errors = payload.errors;
-            appErrors.append($('<ul id="error_list">'));
-            let errorList = $('#error_list');
-            for (let field in errors) {
-                let fieldErrors = errors[field];
-                for (let i = 0; i < fieldErrors.length; i++) {
-                    errorList.append($('<li>').html(fieldErrors[i]))
-                }
-            }
-            break;
-        case 500:
-            return displayError('OOPS! :( Seems like something went wrong. Try again later.');
-    }
-    appErrors.show();
-};
-
-const postSignUp = (body) => {
-    let responseOk = false;
-    let responseStatus = 0;
-    fetch(apiUrl('signup'), {
-        method: "post",
-        body: JSON.stringify(body),
-        headers: {
-            "Content-Type": "application/json"
-        }
-    })
-        .then((response) => {
-            responseOk = response.ok;
-            responseStatus = response.status;
-            return response.json();
-        })
-        .then(data => {
-            if (responseOk) {
-                displaySuccess('Successfully signed up. <a href="signin.html">Sign in</a> to get started!')
-            } else {
-                handleErrors(responseStatus, data);
-                if (responseStatus === 409) {
-                    displayError("A user with the same email address already exists.");
-                } else {
-                    handleErrors(responseStatus, data);
-                }
-            }
-        })
-        .catch((error) => {
-            console.log(error)
-        });
-};
-
-const signUp = (ev) => {
-    ev.preventDefault();
-    let name = getValue('name');
-    let email = getValue('email');
-    let password = getValue('password');
-    // let password_conf = getValue('password_conf');
-    postSignUp({name: name, email: email, password: password});
-};
-
-const postLogin = (body) => {
-    let responseOk = false;
-    let responseStatus = 0;
-    fetch(apiUrl('login'), {
-        method: "post",
-        body: JSON.stringify(body),
-        headers: {
-            "Content-Type": "application/json"
-        }
-    })
-        .then((response) => {
-            responseOk = response.ok;
-            responseStatus = response.status;
-            return response.json();
-        })
-        .then(data => {
-            if (responseOk) {
-                setToken(data.token);
-                window.location = 'entries.html';
-            } else {
-                if (responseStatus === 401) {
-                    displayError("Invalid email or password. Try again.")
-                } else {
-                    handleErrors(responseStatus, data);
-                }
-            }
-        })
-        .catch((error) => {
-            console.log(error)
-        });
-};
-
-const login = (ev) => {
-    ev.preventDefault();
-    let email = getValue('email');
-    let password = getValue('password');
-    postLogin({email: email, password: password})
-};
-
-const requireSignIn = () => {
-    clearToken();
-    window.location = 'signin.html';
-};
-
-const fetchEntries = () => {
-    let responseOk = false;
-    let responseStatus = 0;
-    fetch(apiUrl('entries'), {
-        headers: {
-            "Content-Type": "application/json",
-            "x-access-token": getToken()
-        }
-    })
-        .then((response) => {
-            responseOk = response.ok;
-            responseStatus = response.status;
-            return response.json();
-        })
-        .then(data => {
-            if (responseOk) {
-                let count = data.count;
-                if (count) {
-                    let entries = data.entries;
-                    for (let i = 0; i < count; i++) {
-                        let id = entries[i].id;
-                        let title = entries[i].title;
-                        let created_at = entries[i].created_at;
-                        $('#entries_display').append(
-                            '<div class="entry-card" id="entry_id_' + id + '">' +
-                            '   <div class="row">' +
-                            '       <div class="col-m4">' +
-                            '           <a href="view.html?id=' + id + '" class="btn btn-sm">' + title + '</a>' +
-                            '       </div>' +
-                            '       <div class="col-m4">' + created_at + '</div>' +
-                            '       <div class="col-m4">' +
-                            '           <a href="edit.html?id=' + id + '" class="btn-small">Edit</a>' +
-                            '           <a href onclick="deleteEntry(event, ' + id + ')" class="btn-small">Delete</a>' +
-                            '        </div>' +
-                            '   </div>' +
-                            '</div>'
-                        );
-                    }
-                } else {
-                    displaySuccess("You have no entries.")
-                }
-            } else {
-                if (responseStatus === 401) {
-                    requireSignIn();
-                } else {
-                    handleErrors(responseStatus, data);
-                }
-            }
-        })
-        .catch((error) => {
-            console.log(error)
-        });
-};
-
-const showEntry = (id) => {
-    let responseOk = false;
-    let responseStatus = 0;
-    fetch(apiUrl('entries/' + id), {
-        headers: {
-            "Content-Type": "application/json",
-            "x-access-token": getToken()
-        }
-    })
-        .then((response) => {
-            responseOk = response.ok;
-            responseStatus = response.status;
-            return response.json();
-        })
-        .then(data => {
-                if (responseOk) {
-                    let entry = data.entry;
-                    $('#entry_display').html(
-                        '<div class="panel">' +
-                        '   <div class="panel-body">' +
-                        '       <div class="entry">' +
-                        '           <a href="edit.html?id=' + entry.id + '" class="link-button">Edit</a>' +
-                        '           <h4>Date: ' + entry.created_at + '</h4>' +
-                        '           <h4>Title: ' + entry.title + '</h4>' +
-                        '           <p>' + entry.body + '</p>' +
-                        '           <a href onclick="deleteEntry(event, ' + id + ')" class="link-button">Delete</a>' +
-                        '           </div>' +
-                        '   </div>' +
-                        '</div>'
-                    );
-                } else {
-                    if (responseStatus === 401) {
-                        requireSignIn();
-                    } else {
-                        handleErrors(responseStatus, data);
-                    }
-                }
-            }
-        )
-        .catch((error) => {
-            console.log(error)
-        });
-};
-
-const getUri = (href) => {
-    let location = href ? href : window.location.href;
-    return location.substr(location.lastIndexOf('/') + 1);
-};
-
-const clearMessages = () => {
-    appErrors.hide();
-    appSuccess.hide();
-};
-
-const deleteEntry = (ev, id) => {
-    ev.preventDefault();
-    clearMessages();
-    let responseOk = false;
-    let responseStatus = 0;
-    fetch(apiUrl('entries/' + id), {
-        method: "delete",
-        headers: {
-            "Content-Type": "application/json",
-            "x-access-token": getToken()
-        }
-    })
-        .then((response) => {
-            responseOk = response.ok;
-            responseStatus = response.status;
-            return response.json();
-        })
-        .then(data => {
-                if (responseOk) {
-                    if (getUri().startsWith('view.html')) {
-                        displaySuccess('The entry was deleted successfully');
-                        $('#entry_display').empty();
-                    } else {
-                        $('#entry_id_' +  id).remove();
-                    }
-                } else {
-                    if (responseStatus === 401) {
-                        requireSignIn();
-                    } else {
-                        handleErrors(responseStatus, data);
-                    }
-                }
-            }
-        )
-        .catch((error) => {
-            console.log(error)
-        });
-};
-
-const createEntry = () => {
-    console.log('EXPECT CREATION OF ENTRY');
-};
-
-const showProfile = () => {
-    console.log('SHOW PROFILE');
-};
-
-const loadPage = (uri, url) => {
-    clearMessages();
-    // hide any error or success messages when the user starts typing
-    $('input').keyup(() => clearMessages());
-    const token = getToken();
-    if (uri === 'signup.html' || uri === 'signin.html') {
-        if (token) {
-            window.location = 'entries.html'
-        }
-        // hide logout link
-        $('#logout').hide();
-        return
-    }
-    if (token) {
-        $('a[href^="sign"]').hide();
-        // hide sign in and sign up links
-        switch (uri) {
-            case 'entries.html':
-                fetchEntries();
-                break;
-            case 'create.html':
-                break;
-            case 'profile.html':
-                showProfile();
-                break;
-        }
-        if (uri.startsWith('view.html')) {
-            let id = (new URL(url)).searchParams.get('id');
-            showEntry(id);
-        }
-    } else {
-        window.location = 'signin.html'
-    }
-
-};
-
-(() => {
-    const href = window.location.href;
-    const uri = getUri(href);
-
-    // activate correct link
-    const current = document.querySelector("a[href='" + uri + "']");
-    if (current !== null) {
-        current.className += ' active';
-    }
-
-    // check for auth
-    checkAuth(uri);
-    // a token is available. we are just not sure it's valid
-    loadPage(uri, href);
-})();
+!function(e){var t={};function n(r){if(t[r])return t[r].exports;var i=t[r]={i:r,l:!1,exports:{}};return e[r].call(i.exports,i,i.exports,n),i.l=!0,i.exports}n.m=e,n.c=t,n.d=function(e,t,r){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:r})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var i in e)n.d(r,i,function(t){return e[t]}.bind(null,i));return r},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=6)}([function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.registerLogoutListener=t.requireSignIn=t.checkAuth=t.clearToken=t.setToken=t.getToken=void 0;var r=n(3),i=t.getToken=function(){return localStorage.getItem("auth-token")},o=(t.setToken=function(e){return localStorage.setItem("auth-token",e)},t.clearToken=function(){return localStorage.removeItem("auth-token")});t.checkAuth=function(e){"signup.html"!==e&&"signin.html"!==e&&(i()||(window.location.href="signin.html"))},t.requireSignIn=function(){o(),window.location="signin.html"},t.registerLogoutListener=function(){(0,r.getElement)("#logout").addEventListener("click",function(e){e.preventDefault(),o(),window.location="index.html"})}},function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});var r=$("#app_errors"),i=$("#app_success"),o=(t.clearMessages=function(){r.hide(),i.hide()},t.displaySuccess=function(e){i.empty(),i.append($("<ul>").append($("<li>").html(e))),i.show()},t.displayError=function(e){r.empty(),r.append($("<ul>").append($("<li>").html(e))),r.show()});t.handleErrors=function(e,t){switch(r.empty(),e){case 404:return o("The entry was not found.");case 422:var n=t.errors;r.append($('<ul id="error_list">'));var i=$("#error_list");for(var s in n)for(var a=n[s],l=0;l<a.length;l++)i.append($("<li>").html(a[l]));break;case 500:return o("OOPS! :( Seems like something went wrong. Try again later.")}r.show()}},function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});t.apiUrl=function(e){return"https://mdiary-v1.herokuapp.com/api/v1/"+e}},function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});var r=t.getElement=function(e){var t=e.substring(1,e.length);switch(e.charAt(0)){case"#":return document.getElementById(t);case".":return document.getElementsByClassName(t)}};t.getValue=function(e){var t=r("#"+e);return""===t.value?null:t.value}},function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.loadPage=t.getUri=void 0;var r=n(0),i=n(1),o=n(7),s=n(8),a=n(9),l=n(10);t.getUri=function(e){var t=e||window.location.href;return t.substr(t.lastIndexOf("/")+1)},t.loadPage=function(e,t){(0,i.clearMessages)(),$("input").keyup(function(){return(0,i.clearMessages)()});var n=(0,r.getToken)();if("signup.html"!==e&&"signin.html"!==e)if(n){switch((0,r.registerLogoutListener)(),$('a[href^="sign"]').hide(),e){case"entries.html":(0,a.showEntries)();break;case"create.html":break;case"profile.html":showProfile()}if(e.startsWith("view.html")){var c=new URL(t).searchParams.get("id");(0,l.showEntry)(c)}}else window.location="signin.html";else{switch(n&&(window.location="entries.html"),e){case"signin.html":(0,o.registerLoginListener)();break;case"signup.html":(0,s.registerSignUpListener)()}$("#logout").hide()}}},function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.registerDeleteListener=void 0;var r=n(3),i=n(1),o=n(4),s=n(0),a=n(2);t.registerDeleteListener=function(){(0,i.clearMessages)(),Array.from((0,r.getElement)(".delete")).forEach(function(e){e.addEventListener("click",function(e){e.preventDefault(),function(e){var t=!1,n=0;fetch((0,a.apiUrl)("entries/"+e),{method:"delete",headers:{"Content-Type":"application/json","x-access-token":(0,s.getToken)()}}).then(function(e){return t=e.ok,n=e.status,e.json()}).then(function(r){t?(0,o.getUri)().startsWith("view.html")?((0,i.displaySuccess)("The entry was deleted successfully"),$("#entry_display").empty()):$("#entry_id_"+e).remove():401===n?(0,s.requireSignIn)():(0,i.handleErrors)(n,r)}).catch(function(e){console.log(e)})}(e.target.dataset.entry)})})}},function(e,t,n){"use strict";var r=n(0),i=n(4);!function(){var e=window.location.href,t=(0,i.getUri)(e),n=document.querySelector("a[href='"+t+"']");null!==n&&(n.className+=" active"),(0,r.checkAuth)(t),(0,i.loadPage)(t,e)}()},function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.registerLoginListener=void 0;var r=n(2),i=n(0),o=n(3),s=n(1),a=function(){!function(e){var t=!1,n=0;fetch((0,r.apiUrl)("login"),{method:"post",body:JSON.stringify(e),headers:{"Content-Type":"application/json"}}).then(function(e){return t=e.ok,n=e.status,e.json()}).then(function(e){t?((0,i.setToken)(e.token),window.location="entries.html"):401===n?(0,s.displayError)("Invalid email or password. Try again."):(0,s.handleErrors)(n,e)}).catch(function(e){console.log(e)})}({email:(0,o.getValue)("email"),password:(0,o.getValue)("password")})};t.registerLoginListener=function(){(0,o.getElement)("#login_form").addEventListener("submit",function(e){e.preventDefault(),a()})}},function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.registerSignUpListener=void 0;var r=n(2),i=n(3),o=n(1),s=function(){!function(e){var t=!1,n=0;fetch((0,r.apiUrl)("signup"),{method:"post",body:JSON.stringify(e),headers:{"Content-Type":"application/json"}}).then(function(e){return t=e.ok,n=e.status,e.json()}).then(function(e){t?(0,o.displaySuccess)('Successfully signed up. <a href="signin.html">Sign in</a> to get started!'):((0,o.handleErrors)(n,e),409===n?(0,o.displayError)("A user with the same email address already exists."):(0,o.handleErrors)(n,e))}).catch(function(e){console.log(e)})}({name:(0,i.getValue)("name"),email:(0,i.getValue)("email"),password:(0,i.getValue)("password")})};t.registerSignUpListener=function(){(0,i.getElement)("#sign_up_form").addEventListener("submit",function(e){e.preventDefault(),s()})}},function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.showEntries=void 0;var r=n(2),i=n(0),o=n(1),s=n(5);t.showEntries=function(){var e=!1,t=0;fetch((0,r.apiUrl)("entries"),{headers:{"Content-Type":"application/json","x-access-token":(0,i.getToken)()}}).then(function(n){return e=n.ok,t=n.status,n.json()}).then(function(n){if(e){var r=n.count;if(r){for(var a=n.entries,l=0;l<r;l++){var c=a[l].id,u=a[l].title,d=a[l].created_at;$("#entries_display").append('<div class="entry-card" id="entry_id_'+c+'">   <div class="row">       <div class="col-m4">           <a href="view.html?id='+c+'" class="btn btn-sm">'+u+'</a>       </div>       <div class="col-m4">'+d+'</div>       <div class="col-m4">           <a href="edit.html?id='+c+'" class="btn-small">Edit</a>           <a href data-entry="'+c+'" class="delete btn-small">Delete</a>        </div>   </div></div>')}(0,s.registerDeleteListener)()}else(0,o.displaySuccess)("You have no entries.")}else 401===t?(0,i.requireSignIn)():(0,o.handleErrors)(t,n)}).catch(function(e){console.log(e)})}},function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.showEntry=void 0;var r=n(5),i=n(0),o=n(1),s=n(2);t.showEntry=function(e){var t=!1,n=0;fetch((0,s.apiUrl)("entries/"+e),{headers:{"Content-Type":"application/json","x-access-token":(0,i.getToken)()}}).then(function(e){return t=e.ok,n=e.status,e.json()}).then(function(s){if(t){var a=s.entry;$("#entry_display").html('<div class="panel">   <div class="panel-body">       <div class="entry">           <a href="edit.html?id='+a.id+'" class="link-button">Edit</a>           <h4>Date: '+a.created_at+"</h4>           <h4>Title: "+a.title+"</h4>           <p>"+a.body+'</p>           <a href data-entry="'+e+'" class="delete link-button">Delete</a>           </div>   </div></div>'),(0,r.registerDeleteListener)()}else 401===n?(0,i.requireSignIn)():(0,o.handleErrors)(n,s)}).catch(function(e){console.log(e)})}}]);
