+++ scripts/entries/create.js
@@ -1,0 +1,3 @@
+const createEntry = () => {
+    console.log('CREATE ENTRY');
+};
