+++ scripts/profile.js
@@ -1,0 +1,3 @@
+const showProfile = () => {
+    console.log('SHOW PROFILE');
+};
