+++ signup.html
@@ -19,7 +19,7 @@
         <li><a href="entries.html">My entries</a></li>
         <li><a href="create.html">Write</a></li>
         <li><a href="profile.html">Profile</a></li>
-        <li class="right"><a href onclick="logout(event)" id="logout">Logout</a></li>
+        <li class="right"><a href id="logout">Logout</a></li>
         <li class="right"><a href="signin.html">Sign in</a></li>
         <li class="right"><a href="signup.html">Sign up</a></li>
     </ul>
@@ -32,7 +32,7 @@
             <div class="panel">
                 <div class="panel-heading">Sign up</div>
                 <div class="panel-body">
-                    <form>
+                    <form id="sign_up_form">
                         <div class="row">
                             <div class="col-m4">
                                 <label for="name">Name</label>
@@ -75,7 +75,7 @@
                         </div>
                         <div class="row">
                             <div class="col-m12" style="float: right">
-                                <input type="submit" value="Sign up" onclick="return signUp(event)">
+                                <input type="submit" value="Sign up">
                             </div>
                         </div>
                     </form>
